Added new file with scaling helper functions. Updated recarbonation.
