+++ watertap/examples/chemistry/tests/test_pH_dependent_solubility.py
@@ -131,6 +131,15 @@ from idaes.generic_models.unit_models.equilibrium_reactor import EquilibriumReac
 # Import log10 function from pyomo
 from pyomo.environ import log10
 
+# Import scaling helper functions
+from watertap.examples.chemistry.chem_scaling_utils import (
+    _set_eps_vals,
+    _set_equ_rxn_scaling,
+    _set_mat_bal_scaling_FpcTP,
+    _set_mat_bal_scaling_FTPx,
+    _set_ene_bal_scaling,
+)
+
 __author__ = "Austin Ladshaw"
 
 # Case 1 Config
@@ -293,81 +302,6 @@ case1_log_rxn_config = {
 
 # Get default solver for testing
 solver = get_solver()
-
-## Set of helper functions for scaling
-def _set_eps_vals(rxn_params, rxn_config):
-    # For solubility reactions, have to set the eps value
-    try:
-        for rid in rxn_params.equilibrium_reaction_idx:
-            scale = rxn_config["equilibrium_reactions"][rid]["parameter_data"]["k_eq_ref"][0]
-
-            # NOTE: The solubility_product function STILL has an eps value that we need to set
-            try:
-                # Want to set eps in some fashion similar to this
-                if scale < 1e-16:
-                    rxn_params.component("reaction_"+rid).eps.value = scale*1e-2
-                else:
-                    rxn_params.component("reaction_"+rid).eps.value = 1e-16*1e-2
-            except:
-                pass
-    except:
-        pass
-
-def _set_equ_rxn_scaling(unit, rxn_config):
-    #Add scaling factors for reactions (changes depending on if it is a log form or not)
-    min_scale = 1e-3
-    for i in unit.control_volume.equilibrium_reaction_extent_index:
-        # i[0] = time, i[1] = reaction
-
-        scale = max(min_scale, rxn_config["equilibrium_reactions"][i[1]]["parameter_data"]["k_eq_ref"][0])
-
-        # this may also need to be different for solubility_product
-        if rxn_config["equilibrium_reactions"][i[1]]["equilibrium_form"] == solubility_product:
-            iscale.set_scaling_factor(unit.control_volume.equilibrium_reaction_extent[0.0,i[1]], 1e-1/scale)
-            iscale.constraint_scaling_transform(
-                unit.control_volume.reactions[0.0].equilibrium_constraint[i[1]], 1e-1/scale)
-        else:
-            iscale.set_scaling_factor(unit.control_volume.equilibrium_reaction_extent[0.0,i[1]], 10/scale)
-            iscale.constraint_scaling_transform(
-                unit.control_volume.reactions[0.0].equilibrium_constraint[i[1]], 0.1)
-
-def _set_mat_bal_scaling_FpcTP(unit):
-    # For species
-    min = 1e-3
-    for i in unit.control_volume.properties_out[0.0].mole_frac_phase_comp:
-        # i[0] = phase, i[1] = species
-        if unit.inlet.flow_mol_phase_comp[0, i[0], i[1]].value > min:
-            scale = unit.inlet.flow_mol_phase_comp[0, i[0], i[1]].value
-        else:
-            scale = min
-
-        ## For now, these are both set the same, but may change in testing
-        # Scaling factors for liquid
-        if i[0] == 'Liq':
-            iscale.set_scaling_factor(unit.control_volume.properties_out[0.0].mole_frac_comp[i[1]], 10/scale)
-            iscale.set_scaling_factor(unit.control_volume.properties_out[0.0].mole_frac_phase_comp[i], 10/scale)
-            iscale.set_scaling_factor(unit.control_volume.properties_out[0.0].flow_mol_phase_comp[i], 10/scale)
-            iscale.constraint_scaling_transform(unit.control_volume.material_balances[0.0,i[1]], 10/scale)
-        #Scaling factors for solids
-        else:
-            iscale.set_scaling_factor(unit.control_volume.properties_out[0.0].mole_frac_comp[i[1]], 10/scale)
-            iscale.set_scaling_factor(unit.control_volume.properties_out[0.0].mole_frac_phase_comp[i], 10/scale)
-            iscale.set_scaling_factor(unit.control_volume.properties_out[0.0].flow_mol_phase_comp[i], 10/scale)
-            iscale.constraint_scaling_transform(unit.control_volume.material_balances[0.0,i[1]], 10/scale)
-
-def _set_ene_bal_scaling(unit):
-    max = 1
-    min = 1
-    for phase in unit.control_volume.properties_in[0.0].enth_mol_phase:
-        val = abs(value(unit.control_volume.properties_in[0.0].enth_mol_phase[phase].expr))
-        if val >= max:
-            max = val
-        if val <= min:
-            val = min
-        iscale.set_scaling_factor(unit.control_volume.properties_in[0.0]._enthalpy_flow_term[phase], 10/val)
-        iscale.set_scaling_factor(unit.control_volume.properties_out[0.0]._enthalpy_flow_term[phase], 10/val)
-
-    iscale.constraint_scaling_transform(unit.control_volume.enthalpy_balances[0.0], 10/max)
 
 # Defaults to pH of 7 with no lime added
 def run_case1(xOH=1e-7/55.2, xH=1e-7/55.2, xCaOH2=1e-20, xCa=1e-20,
